+++ kit/src/Runtime/GC.h
@@ -8,34 +8,34 @@
 #define GC_H
 
 #ifdef ENABLE_GC
-extern int time_to_gc;
-extern unsigned int rp_gc_treshold;
-extern unsigned int *stack_bot_gc;
-extern unsigned int alloc_period;
-extern int *data_lab_ptr;
+extern size_t time_to_gc;
+extern size_t rp_gc_treshold;
+extern size_t *stack_bot_gc;
+extern size_t alloc_period;
+extern uintptr_t *data_lab_ptr;
 
-extern unsigned int lobjs_current;
-extern unsigned int lobjs_gc_treshold;
-extern unsigned int lobjs_period;
+extern size_t lobjs_current;
+extern size_t lobjs_gc_treshold;
+extern size_t lobjs_period;
 
-extern int doing_gc;
+extern ssize_t doing_gc;
 #ifdef ENABLE_GEN_GC
-extern int major_p;
-extern int num_gc_major;
+extern ssize_t major_p;
+extern ssize_t num_gc_major;
 #endif
-extern int num_gc;
+extern ssize_t num_gc;
 
-extern int raised_exn_interupt;
-extern int raised_exn_overflow;
+extern ssize_t raised_exn_interupt;
+extern ssize_t raised_exn_overflow;
 
-extern int time_gc_all_ms;
+extern ssize_t time_gc_all_ms;
 
-extern unsigned int *data_begin_addr;
-extern unsigned int *data_end_addr;
+extern size_t *data_begin_addr;
+extern size_t *data_end_addr;
 
-unsigned int size_lobj(unsigned int tag);
+size_t size_lobj(size_t tag);
 
-void gc(unsigned int **sp, unsigned int reg_map);
+void gc(size_t **sp, size_t reg_map);
 
 #endif /*ENABLE_GC*/
